+++ Kernel/Modules/AgentTicketProcess.pm
@@ -260,8 +260,21 @@ sub Run {
         }
     }
 
+    # list only Active processes by default
+    my @ProcessStates = ('Active');
+
+    # set AJAXDialog for proper error responses, screen display and process list
+    $Self->{AJAXDialog} = $Self->{ParamObject}->GetParam( Param => 'AJAXDialog' ) || '';
+
+    # fetch also FadeAway processes to continue working with existing tickets, but not to start new
+    #    ones
+    if ( !$Self->{AJAXDialog} && $Self->{Subaction} ) {
+        push @ProcessStates, 'FadeAway'
+    }
+
+    # get processes
     my $ProcessList = $Self->{ProcessObject}->ProcessList(
-        ProcessState => ['Active'],
+        ProcessState => \@ProcessStates,
         Interface    => ['AgentInterface'],
     );
     my $ProcessEntityID = $Self->{ParamObject}->GetParam( Param => 'ProcessEntityID' );
@@ -272,9 +285,6 @@ sub Run {
             Comment => 'Please contact the admin.',
         );
     }
-
-    # set AJAXDialog for proper error responses and screen display
-    $Self->{AJAXDialog} = $Self->{ParamObject}->GetParam( Param => 'AJAXDialog' ) || '';
 
     # If we have no Subaction or Subaction is 'Create' and submitted ProcessEntityID is invalid
     # Display the ProcessList
@@ -291,6 +301,29 @@ sub Run {
             %Param,
             ProcessList     => $ProcessList,
             ProcessEntityID => $ProcessEntityID
+        );
+    }
+
+    # check if the selected process from the list is valid, prevent tamper with process selection
+    #    list (not existing, invalid an fade away processes must not be able to start a new process
+    #    ticket)
+    elsif (
+        $Self->{Subaction} eq 'DisplayActivityDialogAJAX'
+        && !$ProcessList->{$ProcessEntityID}
+        && $Self->{AJAXDialog}
+        )
+    {
+
+        # translate the error message (as it will be injected in the HTML)
+        my $ErrorMessage
+            = $Self->{LayoutObject}->{LanguageObject}->Get("The selected process is invalid!");
+
+        # return a predefined HTML sctructure as the AJAX call is expecting and HTML response
+        return $Self->{LayoutObject}->Attachment(
+            ContentType => 'text/html; charset=' . $Self->{LayoutObject}->{Charset},
+            Content     => '<div class="ServerError" data-message="' . $ErrorMessage . '"></div>',
+            Type        => 'inline',
+            NoCache     => 1,
         );
     }